refactor(member,auth) : 컨벤션 수정 및 공백 제거
