+++ src/main/java/roomescape/auth/AuthController.java
@@ -1,6 +1,6 @@
 package roomescape.auth;
 
-import jakarta.servlet.http.Cookie;
+
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
