+++ src/main/java/roomescape/member/Member.java
@@ -2,10 +2,10 @@ package roomescape.member;
 
 public class Member {
     private Long id;
-    private String name;
-    private String email;
+    private final String name;
+    private final String email;
     private String password;
-    private String role;
+    private final String role;
 
     public Member(Long id, String name, String email, String role) {
         this.id = id;
